% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topkgraphs.R
\name{topkgraphs}
\alias{topkgraphs}
\title{TopKGraphs}
\usage{
topkgraphs(
  views,
  walk_depth = 50,
  n_iter = 50,
  n_cores = NaN,
  agg_method = "mean"
)
}
\arguments{
\item{views}{List of graphs or adjacency matrices}

\item{walk_depth}{Integer walk length}

\item{n_iter}{Number of walks}

\item{n_cores}{Number of cores (NA = serial)}

\item{agg_method}{Aggregation method}
}
\value{
List with RES, DIST, DIST_raw
}
\description{
Compute node affinities using Jaccard-anchored random walks
}
