% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topkgraphs_walk.R
\name{topkgraphs_walk}
\alias{topkgraphs_walk}
\title{Jaccard-Anchored Random Walk Ranking}
\usage{
topkgraphs_walk(views, adj_list, start_node = 1, walk_depth = 20, n_iter = 20)
}
\arguments{
\item{views}{List of igraph objects or adjacency matrices}

\item{adj_list}{Adjaceny list of each graph}

\item{start_node}{Integer start node}

\item{walk_depth}{Length of each walk}

\item{n_iter}{Number of walks}
}
\value{
A list containing aggregated rankings and scores
}
\description{
Performs multiple biased random walks from a start node and aggregates
the resulting partial rankings.
}
